Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -3,8 +3,8 @@
 #include<string.h>
 
 typedef struct rectangle {
-    int length;
-    int breadth;
+    float length;
+    float breadth;
 } rectangle;
 
 int main(){
@@ -12,7 +12,7 @@ int main(){
     scanf("%d", &a);
     rectangle r[a];
     for(int i = 0; i<a; i++){
-        scanf("%d %d", &r[i].length, &r[i].breadth);
+        scanf("%f %f", &r[i].length, &r[i].breadth);
     }
     for(int i =0;i<a;i++){
         int area = r[i].length * r[i].breadth;
